Updating the regression problem to work with MGOpt infrastructure
